refactor(skill-svc): Layer 1 — extract pure formulas into skill_progression package
Zero-behavior-change extraction of the pure mathematical layer from
skill_progression_service.py into a new app/services/skill_progression/
sub-package.

Moved to _formulas.py (no DB, no model imports):
  - Constants: MIN_SKILL_VALUE, MAX_SKILL_VALUE, MAX_SKILL_CAP, DEFAULT_BASELINE
  - calculate_skill_value_from_placement() — V3 EMA + V2 legacy paths
  - get_skill_tier()

skill_progression_service.py becomes a thin shim that imports and
re-exports all six symbols from the package. All existing callers continue
to resolve through the unchanged import surface with zero modification.

__init__.py re-exports Layer 1 symbols; future layers (config resolution,
DB helpers, EMA engine, views) will be added in subsequent PRs.
